Patch Postgres -  More robust mechanism for getting header (#22)
* More robust mechanism for getting header

the keep-open method needs to access  the header names first before getting chunks of returned content.  The previous method used a SELECT call which returned header-only data.frame for some DBs, but not for POSTGRES, which returned empty content. As a result, ark would cut off table column names for Postgres DBs.

This uses a more generic DBI function to access the header, which preserves headers for Postgres and should work with all DBI connection types.