slightly more robust bulk import
Ultimately this probably needs to be more configurable, in particular, in supporting colClasses specification (or at least a default) and in specifying quote char.  This should be captured from the `...` argument to `unark` and past to bulk.

for example, readr default tsv still uses quotes as escape. (e.g. all taxadb data-raw call write_tsv without specifying `quote=""` as in the tsv standard...).